Update Test for Update delete and update

+++ Service/Services/ClientService.cs
@@ -2,6 +2,7 @@
 using Service.Interface;
 using Service.Mapping;
 using Service.Repository;
+using System.Threading;
 
 namespace Service.Services
 {
@@ -21,7 +22,7 @@ namespace Service.Services
             await _clientRepository.SaveChanges();
         }
 
-        public async Task Update(ClientDTO clientDTO,CancellationToken cancellationToken = default)
+        public async Task Update(ClientDTO clientDTO, CancellationToken cancellationToken = default)
         {
             try
             {
@@ -35,7 +36,8 @@ namespace Service.Services
                 client.GenderId = clientDTO.GenderId;
                 await _clientRepository.Update(client);
                 await _clientRepository.SaveChanges();
-            }catch(Exception ex)
+            }
+            catch (Exception ex)
             {
                 throw;
             }
@@ -44,7 +46,7 @@ namespace Service.Services
         public async Task<ClientDTO> Get(long id, CancellationToken cancellationToken)
         {
 
-            var client=  await _clientRepository.Get(id, cancellationToken);
+            var client = await _clientRepository.Get(id, cancellationToken);
             var clientDTO = ClientMapping.ToClientDTO(client);
             return clientDTO;
         }
@@ -63,12 +65,15 @@ namespace Service.Services
         public async Task<bool> ClientExist(long id, CancellationToken cancellationToken)
 
         {
-           return await _clientRepository.ClientExist(id, cancellationToken);
+            return await _clientRepository.ClientExist(id, cancellationToken);
         }
 
-        public async Task DeleteClient(long id)
+        public async Task DeleteClient(long id, CancellationToken cancellationToken = default)
 
         {
+            var client = await _clientRepository.Get(id, cancellationToken);
+            if (client is null)
+                throw new Exception("Client doesnot exist");
             await _clientRepository.Delete(id);
             await _clientRepository.SaveChanges();
         }
